Remove legacy scan routing and public scan alias
