+++ src/CMakeLists.txt
@@ -17,7 +17,6 @@ set(EXTENSION_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/bigquery_jobs.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bigquery_parser.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bigquery_info.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/bigquery_arrow_scan.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bigquery_geography.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bigquery_secrets.cpp
     PARENT_SCOPE)
