+++ src/bigquery_scan.cpp
@@ -1,363 +1,432 @@
-#include "bigquery_scan.hpp"
-#include "bigquery_arrow_reader.hpp"
-#include "bigquery_arrow_scan.hpp"
-#include "bigquery_client.hpp"
-#include "bigquery_settings.hpp"
-#include "bigquery_sql.hpp"
-#include "bigquery_utils.hpp"
-#include "storage/bigquery_catalog.hpp"
-#include "storage/bigquery_transaction.hpp"
-
-#include "google/cloud/bigquery/storage/v1/arrow.pb.h"
-#include "google/cloud/bigquery/storage/v1/bigquery_read_client.h"
-#include "google/cloud/bigquery/storage/v1/storage.pb.h"
-#include "google/cloud/bigquery/storage/v1/stream.pb.h"
-
-#include <arrow/api.h>
-#include <arrow/c/bridge.h>
-#include <arrow/io/api.h>
-#include <arrow/ipc/api.h>
-#include <arrow/ipc/writer.h>
-
 #include "duckdb.hpp"
 #include "duckdb/common/arrow/arrow.hpp"
 #include "duckdb/common/arrow/arrow_wrapper.hpp"
+#include "duckdb/common/atomic.hpp"
+#include "duckdb/common/limits.hpp"
 #include "duckdb/function/table/arrow.hpp"
 #include "duckdb/main/attached_database.hpp"
 #include "duckdb/main/database_manager.hpp"
 
+#include "bigquery_arrow_reader.hpp"
+#include "bigquery_client.hpp"
+#include "bigquery_scan.hpp"
+#include "bigquery_settings.hpp"
+#include "bigquery_sql.hpp"
+#include "bigquery_utils.hpp"
+
+#include <arrow/c/bridge.h>
+#include <arrow/util/iterator.h>
+#include <unordered_map>
+
 namespace duckdb {
+
+unique_ptr<ArrowArrayStreamWrapper> ProduceArrowScan(const ArrowScanFunctionData &function,
+                                                     const vector<column_t> &column_ids,
+                                                     TableFilterSet *filters);
+
 namespace bigquery {
-
-
-struct BigqueryLocalFunctionState : public LocalTableFunctionState {
-    BigqueryLocalFunctionState() : row_offset(0), done(false) {
-    }
-
-    vector<column_t> column_ids;
-    vector<column_t> column_ids_ranked;
-
-    int row_offset;
-    bool done;
-
-    void ScanNextChunk(DataChunk &output, BigqueryGlobalFunctionState &gstate) {
-        if (current_batch == nullptr || row_offset >= current_batch->num_rows()) {
-            if (!ReadNextBatch()) {
-                // Current stream is exhausted, check if we have more streams to read
-                auto streamidx = gstate.NextStreamIndex();
-                read_stream = arrow_reader->GetStream(streamidx);
-                if (read_stream == nullptr) {
-                    done = true;
-                    return;
-                }
-
-                // Reset the row offset for the new stream
-                rows_read = false;
-                row_offset = 0;
-
-                // Try to read from a new stream
-                if (!ReadNextBatch()) {
-                    done = true;
-                    return;
-                }
-            }
-            row_offset = 0;
-        }
-
-        auto slice = current_batch->Slice(row_offset, STANDARD_VECTOR_SIZE);
-        for (idx_t i = 0; i < output.ColumnCount(); i++) {
-            auto col_idx = column_ids[i];
-            if (COLUMN_IDENTIFIER_ROW_ID == col_idx) {
-                continue;
-            }
-            auto rank_idx = column_ids_ranked[i];
-            auto &out_vec = output.data[i];
-            arrow_reader->ReadColumn(slice->column(rank_idx), out_vec);
-        }
-        output.SetCardinality(slice->num_rows());
-
-        row_offset += slice->num_rows();
-        if (row_offset >= current_batch->num_rows()) {
-            current_batch = nullptr;
-        }
-    }
-
-    shared_ptr<BigqueryArrowReader> arrow_reader;
-    shared_ptr<google::cloud::bigquery::storage::v1::ReadStream> read_stream;
-
-private:
-    bool rows_read = false;
-    google::cloud::v2_38::StreamRange<google::cloud::bigquery::storage::v1::ReadRowsResponse> read_rows;
-    google::cloud::v2_38::StreamRange<google::cloud::bigquery::storage::v1::ReadRowsResponse>::iterator read_rows_it;
-    std::shared_ptr<arrow::RecordBatch> current_batch;
-
-    bool ReadNextBatch() {
-        if (!rows_read) {
-            read_rows = arrow_reader->ReadRows(read_stream->name(), row_offset);
-            read_rows_it = read_rows.begin();
-            rows_read = true;
-        }
-
-        if (read_rows_it == read_rows.end()) {
-            return false;
-        }
-
-        if (!read_rows_it->ok()) {
-            std::cerr << "Error reading rows: " << read_rows_it->status() << std::endl;
-            done = true;
-            return false;
-        }
-        current_batch = arrow_reader->ReadBatch(read_rows_it->value().arrow_record_batch());
-        read_rows_it++;
-        return true;
-    }
-};
-
-unique_ptr<FunctionData> BigqueryLegacyScanFunction::BigqueryLegacyScanBind(ClientContext &context,
-                                                                            TableFunctionBindInput &input,
-                                                                            vector<LogicalType> &return_types,
-                                                                            vector<string> &names) {
-    auto table_string = input.inputs[0].GetValue<string>();
-    auto table_ref = BigqueryUtils::ParseTableString(table_string);
-    if (!table_ref.HasDatasetId() || !table_ref.HasTableId()) {
-        throw ParserException("Invalid table string: %s", table_string);
-    }
-
-    // Parse named parameters using centralized function
-    auto params = BigQueryCommonParameters::ParseFromNamedParameters(input.named_parameters);
-
-    auto result = make_uniq<BigqueryLegacyScanBindData>();
-    result->table_ref = table_ref;
-    result->filter_condition = params.filter;
-    result->config = BigqueryConfig(table_ref.project_id)
-                         .SetDatasetId(table_ref.dataset_id)
-                         .SetBillingProjectId(params.billing_project)
-                         .SetApiEndpoint(params.api_endpoint)
-                         .SetGrpcEndpoint(params.grpc_endpoint);
-    result->bq_client = make_shared_ptr<BigqueryClient>(context, result->config);
-
-    ColumnList columns;
-    vector<unique_ptr<Constraint>> constraints;
-
-    string filter_cond = "";
-    if (!params.filter.empty()) {
-        filter_cond = params.filter;
-    }
-
-    auto arrow_reader = result->bq_client->CreateArrowReader(table_ref, 1, std::vector<string>(), filter_cond);
-    arrow_reader->MapTableInfo(columns, constraints);
-
-    for (auto &column : columns.Logical()) {
-        names.push_back(column.GetName());
-        return_types.push_back(column.GetType());
-    }
-    if (names.empty()) {
-        auto table_ref = arrow_reader->GetTableRef();
-        throw std::runtime_error("no columns for table " + table_ref.table_id);
-    }
-
-    for (const auto &column : columns.Logical()) {
-        if (BigqueryUtils::IsGeometryType(column.GetType())) {
-            throw BinderException("BigQuery GEOGRAPHY columns are not supported in legacy scan. "
-                                  "Please set bq_use_legacy_scan=false (recommended).");
-        }
-    }
-
-    // TODO GetMaxRowId
-    result->estimated_row_count = idx_t(arrow_reader->GetEstimatedRowCount());
-    result->names = names;
-    result->types = return_types;
-    return std::move(result);
-}
 
 unique_ptr<FunctionData> BigqueryScanFunction::BigqueryScanBind(ClientContext &context,
                                                                 TableFunctionBindInput &input,
                                                                 vector<LogicalType> &return_types,
                                                                 vector<string> &names) {
-    // Parse named parameters to check use_legacy_scan
+    if (input.inputs.empty()) {
+        throw BinderException("bigquery_scan: table name must be provided");
+    }
+
+    auto table_string = input.inputs[0].GetValue<string>();
+    auto table_ref = BigqueryUtils::ParseTableString(table_string);
+    if (!table_ref.HasDatasetId() || !table_ref.HasTableId()) {
+        throw BinderException("Invalid table string: %s", table_string);
+    }
+
     auto params = BigQueryCommonParameters::ParseFromNamedParameters(input.named_parameters);
 
-    if (params.use_legacy_scan) {
-        return BigqueryLegacyScanFunction::BigqueryLegacyScanBind(context, input, return_types, names);
+    auto bind_data = make_uniq<BigqueryScanBindData>();
+    bind_data->table_ref = table_ref;
+    bind_data->filter_condition = params.filter;
+    bind_data->bq_config = BigqueryConfig(table_ref.project_id)
+                               .SetDatasetId(table_ref.dataset_id)
+                               .SetBillingProjectId(params.billing_project)
+                               .SetApiEndpoint(params.api_endpoint)
+                               .SetGrpcEndpoint(params.grpc_endpoint);
+    bind_data->bq_client = make_shared_ptr<BigqueryClient>(context, bind_data->bq_config);
+
+    ColumnList columns;
+    vector<unique_ptr<Constraint>> constraints;
+    bind_data->bq_client->GetTableInfo(table_ref.dataset_id, table_ref.table_id, columns, constraints);
+
+    auto arrow_schema_ptr = BigqueryUtils::BuildArrowSchema(columns);
+    auto status = arrow::ExportSchema(*std::move(arrow_schema_ptr), &bind_data->schema_root.arrow_schema);
+    if (!status.ok()) {
+        throw BinderException("Arrow schema export failed: " + status.ToString());
+    }
+
+    vector<LogicalType> mapped_bq_types;
+    BigqueryUtils::PopulateAndMapArrowTableTypes(context,
+                                                 bind_data->arrow_table,
+                                                 bind_data->schema_root,
+                                                 names,
+                                                 return_types,
+                                                 mapped_bq_types,
+                                                 &columns);
+
+    bind_data->names = names;
+    bind_data->all_types = return_types;
+    if (!mapped_bq_types.empty()) {
+        bind_data->mapped_bq_types = std::move(mapped_bq_types);
+        bind_data->requires_cast = true;
     } else {
-        return BigqueryArrowScanFunction::BigqueryArrowScanBind(context, input, return_types, names);
-    }
-}
-
-unique_ptr<GlobalTableFunctionState> BigqueryLegacyScanFunction::BigqueryLegacyScanInitGlobalState(
-    ClientContext &context,
-    TableFunctionInitInput &input) {
-    auto &bind_data = input.bind_data->Cast<BigqueryLegacyScanBindData>();
-
-    // selected fields
-    vector<string> selected_fields;
-    for (auto &column_id : input.column_ids) {
-        if (COLUMN_IDENTIFIER_ROW_ID == column_id) {
-            continue;
-        }
-        selected_fields.push_back(bind_data.names[column_id]);
+        bind_data->requires_cast = false;
     }
 
-    // filters
-    bool enable_filter_pushdown = BigquerySettings::ExperimentalFilterPushdown();
-    string filter_string;
-    auto filters = input.filters;
-    if (!bind_data.filter_condition.empty()) {
-        if (!filter_string.empty()) {
-            filter_string += " AND ";
-        }
-        filter_string += bind_data.filter_condition;
-    } else if (enable_filter_pushdown && filters && !filters->filters.empty()) {
-        for (auto &filter : filters->filters) {
-            if (!filter_string.empty()) {
-                filter_string += " AND ";
-            }
-            string column_name = selected_fields[filter.first];
-            auto &filter_cond = *filter.second;
-            filter_string += BigquerySQL::TransformFilter(column_name, filter_cond);
-        }
-    }
-
-    // when preserve_insertion_order=FALSE, we can use multiple streams for parallelization; defaults to maximum_threads
-    // when preserve_insertion_order=TRUE, we use only 1 stream as there won't be any parallelization from DuckDB
-    idx_t k_max_read_streams = BigquerySettings::GetMaxReadStreams(context);
-    auto arrow_reader =
-        bind_data.bq_client->CreateArrowReader(bind_data.table_ref, k_max_read_streams, selected_fields, filter_string);
-
-    auto &mutable_bind_data = input.bind_data->CastNoConst<BigqueryLegacyScanBindData>();
-    mutable_bind_data.estimated_row_count = idx_t(arrow_reader->GetEstimatedRowCount());
-    auto result = make_uniq<BigqueryGlobalFunctionState>(arrow_reader, k_max_read_streams);
-    return std::move(result);
-}
-
-unique_ptr<LocalTableFunctionState> BigqueryLegacyScanFunction::BigqueryLegacyScanInitLocalState(
-    ExecutionContext &context,
-    TableFunctionInitInput &input,
-    GlobalTableFunctionState *global_state) {
-    auto &gstate = global_state->Cast<BigqueryGlobalFunctionState>();
-    auto lstate = make_uniq<BigqueryLocalFunctionState>();
-    lstate->arrow_reader = gstate.arrow_reader;
-
-    lstate->column_ids = input.column_ids;
-    lstate->column_ids_ranked = CalculateRanks(input.column_ids);
-
-    auto streamidx = gstate.NextStreamIndex();
-    lstate->read_stream = gstate.arrow_reader->GetStream(streamidx);
-    if (lstate->read_stream == nullptr) {
-        lstate->done = true;
-    }
-    return std::move(lstate);
-}
-
-void BigqueryLegacyScanFunction::BigqueryLegacyScanExecute(ClientContext &context,
-                                                           TableFunctionInput &data,
-                                                           DataChunk &output) {
-    auto &lstate = data.local_state->Cast<BigqueryLocalFunctionState>();
-    auto &gstate = data.global_state->Cast<BigqueryGlobalFunctionState>();
-    if (lstate.done) {
-        return;
-    }
-    output.Reset();
-    lstate.ScanNextChunk(output, gstate);
-    lock_guard<mutex> glock(gstate.lock);
-    gstate.position += output.size();
+    return std::move(bind_data);
 }
 
 unique_ptr<GlobalTableFunctionState> BigqueryScanFunction::BigqueryScanInitGlobalState(ClientContext &context,
-                                                                                       TableFunctionInitInput &input) {
-    // Check bind data type and forward to appropriate implementation
-    if (dynamic_cast<const BigqueryArrowScanBindData *>(input.bind_data.get())) {
-        return BigqueryArrowScanFunction::BigqueryArrowScanInitGlobal(context, input);
-    } else {
-        return BigqueryLegacyScanFunction::BigqueryLegacyScanInitGlobalState(context, input);
+                                                                                        TableFunctionInitInput &input) {
+    auto &bind_data = input.bind_data->CastNoConst<BigqueryScanBindData>();
+
+    vector<string> selected_fields;
+    selected_fields.reserve(input.column_ids.size());
+    for (auto col_id : input.column_ids) {
+        if (col_id == COLUMN_IDENTIFIER_ROW_ID || col_id < 0) {
+            continue;
+        }
+        selected_fields.emplace_back(bind_data.names[col_id]);
     }
+
+    string filter_string = bind_data.filter_condition;
+    if (BigquerySettings::ExperimentalFilterPushdown() && filter_string.empty() && input.filters &&
+        !input.filters->filters.empty()) {
+        for (auto &filter : input.filters->filters) {
+            column_t logical_col = input.column_ids[filter.first];
+            if (logical_col == COLUMN_IDENTIFIER_ROW_ID || logical_col < 0) {
+                throw InvalidInputException("ROWID cannot be referenced in a WHERE clause for BigQuery tables");
+            }
+
+            if (!filter_string.empty()) {
+                filter_string += " AND ";
+            }
+            const string &column_name = bind_data.names[logical_col];
+            filter_string += BigquerySQL::TransformFilter(column_name, *filter.second);
+        }
+    }
+
+    idx_t max_read_streams = BigquerySettings::GetMaxReadStreams(context);
+    auto bq_arrow_reader = bind_data.bq_client->CreateArrowReader(bind_data.table_ref,
+                                                                  max_read_streams,
+                                                                  selected_fields,
+                                                                  filter_string);
+
+    auto factory = make_shared_ptr<BigqueryStreamFactory>(bq_arrow_reader);
+    auto factory_dependency = bind_data.GetFactoryDependency();
+    if (!factory_dependency) {
+        throw InternalException("Factory dependency not initialized");
+    }
+    factory_dependency->factory = factory;
+    bind_data.stream_factory_ptr = reinterpret_cast<uintptr_t>(factory.get());
+
+    auto gstate = make_uniq<BigqueryScanGlobalState>();
+    gstate->max_threads = max_read_streams;
+    bind_data.estimated_row_count = bq_arrow_reader->GetEstimatedRowCount();
+
+    auto arrow_schema = bq_arrow_reader->GetSchema();
+    const idx_t phys_col_count = arrow_schema->num_fields();
+
+    unordered_map<string, column_t> name_to_bq_physical;
+    name_to_bq_physical.reserve(bind_data.names.size());
+    for (idx_t i = 0; i < bind_data.names.size(); ++i) {
+        name_to_bq_physical.emplace(bind_data.names[i], i);
+    }
+
+    gstate->scanned_types.clear();
+    gstate->scanned_types.reserve(phys_col_count);
+    for (idx_t phys_idx = 0; phys_idx < phys_col_count; ++phys_idx) {
+        const string &field_name = arrow_schema->field(phys_idx)->name();
+        auto col_id = name_to_bq_physical.at(field_name);
+        if (bind_data.mapped_bq_types.empty()) {
+            gstate->scanned_types.emplace_back(bind_data.all_types[col_id]);
+        } else {
+            gstate->scanned_types.emplace_back(bind_data.mapped_bq_types[col_id]);
+        }
+    }
+
+    vector<idx_t> column_physical_positions;
+    column_physical_positions.reserve(input.column_ids.size());
+    const idx_t invalid_phys_idx = NumericLimits<idx_t>::Maximum();
+    bool contains_rowid = false;
+    for (idx_t col_pos = 0; col_pos < input.column_ids.size(); ++col_pos) {
+        idx_t col_id = input.column_ids[col_pos];
+        if (col_id == COLUMN_IDENTIFIER_ROW_ID || col_id < 0) {
+            contains_rowid = true;
+            column_physical_positions.push_back(invalid_phys_idx);
+            continue;
+        }
+        const string &name = bind_data.names[col_id];
+        idx_t phys_idx = static_cast<idx_t>(arrow_schema->GetFieldIndex(name));
+        if (phys_idx == static_cast<idx_t>(-1)) {
+            throw InternalException("Column '" + name + "' not found in Arrow schema");
+        }
+        column_physical_positions.push_back(phys_idx);
+    }
+
+    bool requires_physical_reorder = false;
+    vector<idx_t> projection_mapping;
+    if (!input.projection_ids.empty()) {
+        projection_mapping.reserve(input.projection_ids.size());
+        for (idx_t out_idx = 0; out_idx < input.projection_ids.size(); ++out_idx) {
+            idx_t col_pos = input.projection_ids[out_idx];
+            D_ASSERT(col_pos < column_physical_positions.size());
+            idx_t phys_idx = column_physical_positions[col_pos];
+            if (phys_idx == invalid_phys_idx) {
+                contains_rowid = true;
+                requires_physical_reorder = true;
+                break;
+            }
+            projection_mapping.push_back(phys_idx);
+            if (phys_idx != col_pos) {
+                requires_physical_reorder = true;
+            }
+        }
+        if (!contains_rowid) {
+            gstate->projection_ids = std::move(projection_mapping);
+        } else {
+            gstate->projection_ids.clear();
+        }
+    } else {
+        projection_mapping.reserve(column_physical_positions.size());
+        for (idx_t out_idx = 0; out_idx < column_physical_positions.size(); ++out_idx) {
+            idx_t phys_idx = column_physical_positions[out_idx];
+            if (phys_idx == invalid_phys_idx) {
+                contains_rowid = true;
+                requires_physical_reorder = true;
+                break;
+            }
+            projection_mapping.push_back(phys_idx);
+            if (phys_idx != out_idx && phys_idx != COLUMN_IDENTIFIER_ROW_ID && phys_idx >= 0) {
+                requires_physical_reorder = true;
+            }
+        }
+        if (!contains_rowid && requires_physical_reorder) {
+            gstate->projection_ids = std::move(projection_mapping);
+        } else {
+            gstate->projection_ids.clear();
+        }
+    }
+
+    gstate->stream = ::duckdb::ProduceArrowScan(bind_data, input.column_ids, input.filters.get());
+    return std::move(gstate);
 }
 
 unique_ptr<LocalTableFunctionState> BigqueryScanFunction::BigqueryScanInitLocalState(
     ExecutionContext &context,
     TableFunctionInitInput &input,
-    GlobalTableFunctionState *global_state) {
-    // Check bind data type and forward to appropriate implementation
-    if (dynamic_cast<const BigqueryArrowScanBindData *>(input.bind_data.get())) {
-        return BigqueryArrowScanFunction::BigqueryArrowScanInitLocal(context, input, global_state);
-    } else {
-        return BigqueryLegacyScanFunction::BigqueryLegacyScanInitLocalState(context, input, global_state);
+    GlobalTableFunctionState *global_state_p) {
+    auto &client_context = context.client;
+    auto &bind_data = input.bind_data->CastNoConst<BigqueryScanBindData>();
+    auto &global_state = global_state_p->Cast<ArrowScanGlobalState>();
+
+    auto current_chunk = make_uniq<ArrowArrayWrapper>();
+    auto result = make_uniq<BigqueryScanLocalState>(std::move(current_chunk), client_context);
+
+    auto sorted_column_ids = input.column_ids;
+    std::sort(sorted_column_ids.begin(), sorted_column_ids.end());
+
+    result->column_ids = sorted_column_ids;
+    result->filters = input.filters.get();
+    if (!bind_data.projection_pushdown_enabled) {
+        result->column_ids.clear();
+    } else if (!input.projection_ids.empty() || bind_data.requires_cast || !global_state.projection_ids.empty()) {
+        auto &asgs = global_state_p->Cast<ArrowScanGlobalState>();
+        result->all_columns.Initialize(client_context, asgs.scanned_types);
     }
+    if (!ArrowTableFunction::ArrowScanParallelStateNext(client_context, input.bind_data.get(), *result, global_state)) {
+        return nullptr;
+    }
+
+    return std::move(result);
 }
 
-void BigqueryScanFunction::BigqueryScanExecute(ClientContext &context, TableFunctionInput &data, DataChunk &output) {
-    // Check bind data type and forward to appropriate implementation
-    if (dynamic_cast<const BigqueryArrowScanBindData *>(data.bind_data.get())) {
-        BigqueryArrowScanFunction::BigqueryArrowScanExecute(context, data, output);
-    } else {
-        BigqueryLegacyScanFunction::BigqueryLegacyScanExecute(context, data, output);
+void BigqueryScanFunction::BigqueryScanExecute(ClientContext &ctx, TableFunctionInput &data_p, DataChunk &output) {
+    if (!data_p.local_state) {
+        return;
     }
+
+    auto &data = data_p.bind_data->CastNoConst<BigqueryScanBindData>();
+    auto &state = data_p.local_state->Cast<BigqueryScanLocalState>();
+    auto &gstate = data_p.global_state->Cast<BigqueryScanGlobalState>();
+
+    if (state.chunk_offset >= static_cast<idx_t>(state.chunk->arrow_array.length)) {
+        if (!ArrowTableFunction::ArrowScanParallelStateNext(ctx, data_p.bind_data.get(), state, gstate)) {
+            return;
+        }
+    }
+
+    auto output_size =
+        MinValue<idx_t>(STANDARD_VECTOR_SIZE, NumericCast<idx_t>(state.chunk->arrow_array.length) - state.chunk_offset);
+    data.lines_read += output_size;
+
+    auto ensure_column_id_coverage = [&](DataChunk &chunk) {
+        if (!state.column_ids.empty() && state.column_ids.size() != chunk.ColumnCount()) {
+            state.column_ids.clear();
+        }
+    };
+
+    bool only_rowid_columns = !state.column_ids.empty();
+    for (auto col_id : state.column_ids) {
+        if (col_id != COLUMN_IDENTIFIER_ROW_ID && col_id >= 0) {
+            only_rowid_columns = false;
+            break;
+        }
+    }
+
+    if (gstate.CanRemoveFilterColumns()) {
+        state.all_columns.Reset();
+        state.all_columns.SetCardinality(output_size);
+
+        ensure_column_id_coverage(state.all_columns);
+        bool arrow_scan_is_projected = !state.column_ids.empty() && state.column_ids.size() < data.names.size();
+        ArrowTableFunction::ArrowToDuckDB(state,
+                                          data.arrow_table.GetColumns(),
+                                          state.all_columns,
+                                          arrow_scan_is_projected,
+                                          COLUMN_IDENTIFIER_ROW_ID);
+
+        bool geometry_cast_needed = false;
+        if (!data.requires_cast) {
+            if (gstate.projection_ids.empty()) {
+                for (idx_t i = 0; i < output.ColumnCount(); i++) {
+                    auto &src_vec = state.all_columns.data[i];
+                    auto &dst_type = output.data[i].GetType();
+                    if (dst_type.id() == LogicalTypeId::GEOMETRY && src_vec.GetType().id() == LogicalTypeId::VARCHAR) {
+                        geometry_cast_needed = true;
+                        break;
+                    }
+                }
+            } else {
+                for (idx_t i = 0; i < output.ColumnCount(); i++) {
+                    auto proj_id = gstate.projection_ids[i];
+                    auto &src_vec = state.all_columns.data[proj_id];
+                    auto &dst_type = output.data[i].GetType();
+                    if (dst_type.id() == LogicalTypeId::GEOMETRY && src_vec.GetType().id() == LogicalTypeId::VARCHAR) {
+                        geometry_cast_needed = true;
+                        break;
+                    }
+                }
+            }
+        }
+
+        bool do_cast = data.requires_cast || geometry_cast_needed;
+        if (!do_cast) {
+            output.ReferenceColumns(state.all_columns, gstate.projection_ids);
+        } else {
+            for (idx_t col_idx = 0; col_idx < output.ColumnCount(); col_idx++) {
+                auto proj_id = gstate.projection_ids[col_idx];
+                VectorOperations::Cast(ctx,
+                                       state.all_columns.data[proj_id],
+                                       output.data[col_idx],
+                                       state.all_columns.size());
+            }
+        }
+    } else {
+        output.SetCardinality(output_size);
+
+        bool geometry_cast_needed = false;
+        if (!data.requires_cast) {
+            for (idx_t i = 0; i < output.ColumnCount(); i++) {
+                auto &dst_type = output.data[i].GetType();
+                const LogicalType &src_type = gstate.scanned_types[i];
+                if (dst_type.id() == LogicalTypeId::GEOMETRY && src_type.id() == LogicalTypeId::VARCHAR) {
+                    geometry_cast_needed = true;
+                    break;
+                }
+            }
+        }
+
+        bool do_cast = data.requires_cast || geometry_cast_needed;
+        if (!do_cast) {
+            ensure_column_id_coverage(output);
+            bool arrow_scan_is_projected = !state.column_ids.empty() && state.column_ids.size() < data.names.size();
+            ArrowTableFunction::ArrowToDuckDB(state,
+                                              data.arrow_table.GetColumns(),
+                                              output,
+                                              arrow_scan_is_projected,
+                                              COLUMN_IDENTIFIER_ROW_ID);
+        } else {
+            if (only_rowid_columns) {
+                for (idx_t col_idx = 0; col_idx < output.ColumnCount(); col_idx++) {
+                    output.data[col_idx].Sequence(NumericCast<int64_t>(gstate.position.load()), 1, output_size);
+                }
+                output.SetCardinality(output_size);
+                output.Verify();
+                state.chunk_offset += output.size();
+
+                lock_guard<mutex> glock(gstate.lock);
+                gstate.position += output.size();
+                return;
+            }
+
+            state.all_columns.Reset();
+            state.all_columns.SetCardinality(output_size);
+            ensure_column_id_coverage(state.all_columns);
+            bool arrow_scan_is_projected = !state.column_ids.empty() && state.column_ids.size() < data.names.size();
+            ArrowTableFunction::ArrowToDuckDB(state,
+                                              data.arrow_table.GetColumns(),
+                                              state.all_columns,
+                                              arrow_scan_is_projected,
+                                              COLUMN_IDENTIFIER_ROW_ID);
+            for (idx_t col_idx = 0; col_idx < output.ColumnCount(); col_idx++) {
+                VectorOperations::Cast(ctx,
+                                       state.all_columns.data[col_idx],
+                                       output.data[col_idx],
+                                       state.all_columns.size());
+            }
+        }
+    }
+
+    output.SetCardinality(output_size);
+    output.Verify();
+    state.chunk_offset += output.size();
+
+    lock_guard<mutex> glock(gstate.lock);
+    gstate.position += output.size();
 }
 
 static InsertionOrderPreservingMap<string> BigqueryScanToString(TableFunctionToStringInput &input) {
     D_ASSERT(input.bind_data);
-    InsertionOrderPreservingMap<string> result;
 
-    // Check bind data type and forward to appropriate implementation
-    if (auto *arrow_bind_data = dynamic_cast<const BigqueryArrowScanBindData *>(input.bind_data.get())) {
-        result["Table"] = arrow_bind_data->TableString();
-    } else {
-        auto &bind_data = input.bind_data->Cast<BigqueryLegacyScanBindData>();
-        result["Table"] = bind_data.TableString();
-    }
+    InsertionOrderPreservingMap<string> result;
+    auto &bind_data = input.bind_data->Cast<BigqueryScanBindData>();
+    result["Table"] = bind_data.TableString();
     return result;
 }
 
-unique_ptr<NodeStatistics> BigqueryScanCardinality(ClientContext &context, const FunctionData *bind_data_p) {
-    // Check bind data type and forward to appropriate implementation
-    if (auto *arrow_bind_data = dynamic_cast<const BigqueryArrowScanBindData *>(bind_data_p)) {
-        return make_uniq<NodeStatistics>(arrow_bind_data->estimated_row_count, arrow_bind_data->estimated_row_count);
-    } else {
-        auto &bind_data = bind_data_p->Cast<BigqueryLegacyScanBindData>();
-        return make_uniq<NodeStatistics>(bind_data.estimated_row_count, bind_data.estimated_row_count);
-    }
-}
-
-double BigqueryScanProgress(ClientContext &context,
-                            const FunctionData *bind_data_p,
-                            const GlobalTableFunctionState *global_state) {
-    // Check bind data type and forward to appropriate implementation
-    if (auto *arrow_bind_data = dynamic_cast<const BigqueryArrowScanBindData *>(bind_data_p)) {
-        auto &gstate = global_state->Cast<BigqueryArrowScanGlobalState>();
-        double progress = 0.0;
-        if (arrow_bind_data->estimated_row_count > 0) {
-            lock_guard<mutex> glock(gstate.lock);
-            progress = 100.0 * double(gstate.position) / double(arrow_bind_data->estimated_row_count);
-        }
-        return MinValue<double>(100, progress);
-    } else {
-        auto &bind_data = bind_data_p->Cast<BigqueryLegacyScanBindData>();
-        auto &gstate = global_state->Cast<BigqueryGlobalFunctionState>();
-        double progress = 0.0;
-        if (bind_data.estimated_row_count > 0) {
-            lock_guard<mutex> glock(gstate.lock);
-            progress = 100.0 * double(gstate.position) / double(bind_data.estimated_row_count);
-        }
-        return MinValue<double>(100, progress);
-    }
-}
-
 static BindInfo BigqueryScanGetBindInfo(const optional_ptr<FunctionData> bind_data_p) {
-    // Try to cast to BigqueryArrowScanBindData first (optimized scan)
-    if (auto *arrow_bind_data = dynamic_cast<const BigqueryArrowScanBindData *>(bind_data_p.get())) {
-        BindInfo info(ScanType::EXTERNAL);
-        if (arrow_bind_data->bq_table_entry) {
-            info.table = arrow_bind_data->bq_table_entry.get_mutable();
-        }
-        return info;
-    } else {
-        auto &bind_data = bind_data_p->Cast<BigqueryLegacyScanBindData>();
-        BindInfo info(ScanType::EXTERNAL);
-        if (bind_data.bq_table_entry) {
-            info.table = bind_data.bq_table_entry.get_mutable();
-        }
-        return info;
+    auto &bind_data = bind_data_p->Cast<BigqueryScanBindData>();
+    BindInfo info(ScanType::EXTERNAL);
+    if (bind_data.bq_table_entry) {
+        info.table = bind_data.bq_table_entry.get_mutable();
     }
+    return info;
+}
+
+static unique_ptr<NodeStatistics> BigqueryScanCardinality(ClientContext &, const FunctionData *bind_data_p) {
+    auto &bind_data = bind_data_p->Cast<BigqueryScanBindData>();
+    return make_uniq<NodeStatistics>(bind_data.estimated_row_count, bind_data.estimated_row_count);
+}
+
+static double BigqueryScanProgress(ClientContext &,
+                                   const FunctionData *bind_data_p,
+                                   const GlobalTableFunctionState *global_state) {
+    auto &bind_data = bind_data_p->Cast<BigqueryScanBindData>();
+    auto &gstate = global_state->Cast<BigqueryScanGlobalState>();
+    double progress = 0.0;
+    if (bind_data.estimated_row_count > 0) {
+        lock_guard<mutex> glock(gstate.lock);
+        progress = 100.0 * double(gstate.position) / double(bind_data.estimated_row_count);
+    }
+    return MinValue<double>(100, progress);
 }
 
 BigqueryScanFunction::BigqueryScanFunction()
@@ -367,19 +436,19 @@ BigqueryScanFunction::BigqueryScanFunction()
                     BigqueryScanFunction::BigqueryScanBind,
                     BigqueryScanFunction::BigqueryScanInitGlobalState,
                     BigqueryScanFunction::BigqueryScanInitLocalState) {
+    projection_pushdown = true;
+    filter_pushdown = true;
+    filter_prune = true;
+
     to_string = BigqueryScanToString;
     cardinality = BigqueryScanCardinality;
     table_scan_progress = BigqueryScanProgress;
     get_bind_info = BigqueryScanGetBindInfo;
 
-    projection_pushdown = true;
-    filter_pushdown = true;
-
     named_parameters["billing_project"] = LogicalType::VARCHAR;
     named_parameters["api_endpoint"] = LogicalType::VARCHAR;
     named_parameters["grpc_endpoint"] = LogicalType::VARCHAR;
     named_parameters["filter"] = LogicalType::VARCHAR;
-    named_parameters["use_legacy_scan"] = LogicalType::BOOLEAN;
 }
 
 } // namespace bigquery
