+++ src/include/bigquery_settings.hpp
@@ -225,24 +225,6 @@ public:
         }
     }
 
-    static bool &UseLegacyScan() {
-        static bool BIGQUERY_USE_LEGACY_SCAN = false;
-        return BIGQUERY_USE_LEGACY_SCAN;
-    }
-
-    static void SetUseLegacyScan(ClientContext &context, SetScope scope, Value &parameter) {
-        UseLegacyScan() = BooleanValue::Get(parameter);
-    }
-
-    // Deprecated setting for backward compatibility
-    static void SetExperimentalIncubatingScan(ClientContext &context, SetScope scope, Value &parameter) {
-        bool value = BooleanValue::Get(parameter);
-        UseLegacyScan() = !value;
-
-        printf("WARNING: bq_experimental_use_incubating_scan is deprecated. "
-               "Please use bq_use_legacy_scan=%s instead.\n",
-               value ? "false" : "true");
-    }
 };
